Refactor: Reservation, Inventory 를 구분하는 패키지 추
